Split widgets into separate modules.
